backend/compliance_matcher/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==1.10.13
google-generativeai==0.3.1